// src/lib/sanity/types.ts
export type Workout = {
  _id: string;
  _type: "workout";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  userId?: string;
  dateTime?: string;
  duration?: number;
  exercises?: Array<{
    exercise?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "exercise";
    };
    sets?: Array<{
      reps?: number;
      weight?: number;
      weightUnit?: "lbs" | "kg";
      _type: "set";
      _key: string;
    }>;
    _type: "exerciseSet";
    _key: string;
  }>;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type GetWorkoutsQueryResult = Array<{
  _id: string;
  date: null;
  duration: number | null;
  exercises: Array<null> | null;
}>;